Initial conversion of project to typescript
